WIP: Weapon Plugin
